feat: Rendezvous hashing; choice of algorithm; use u64 not strings
This is a squash of a series of commits from a branch where I did
a lot of tinkering then decided to land it all.

First, there's a rendezvous hash variation called `Recognizer` that is
preferable to consistent hash rings in most use cases. The API for it
is almost identical; it exposes a node "weight" that biases items toward
that node. This is useful if you have a cache host with lots of
extra space for whatever reason.

Made the hash algorithm choice a crate feature. The default is blake3
still, but you can pick one of xxhash, fnv, blake2, or sha2 if you
prefer. It's more important to choose a fast hash than to choose one
with cryptographic properties.

Switched to using numbers instead of strings for hash keys. This
improves performance considerably, though I do not like the
thousand-node numbers. Will I ever run 1000 of these? Well, will
I ever run 10 of these? The answer is likely the same for both,
but the point of the exercise is to do it right.